Ajusta módulo 'account_bank_statement_import' para utilizar novamente
- Ajuste de espaçamento em código.
- Ajusta tamanho máximo do form utilizado para conciliação bancária.
- Ajusta estrutura dos arquivos e remove arquivos inutilziados
  - `__init__` e `__manifest__` ajustados.
