+++ addons/account_bank_statement_import/__init__.py
@@ -1,6 +1,4 @@
 # -*- encoding: utf-8 -*-
 
-# from . import account_bank_statement_import
-# from . import account_journal
-
-# from . import wizard
+from . import wizard
+from . import models
