--- addons/account_bank_statement_import/account_import_tip_data.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <data>
-
-    </data>
-</odoo>
