+++ addons/account_bank_statement_import/models/__init__.py
@@ -1,0 +1,3 @@
+from . import account_journal
+from . import account_bank_statement_import
+from . import account_bank_statement_line
