+++ addons/account_bank_statement_import/models/account_bank_statement_import.py
@@ -6,20 +6,6 @@ from odoo import api, fields, models, _
 from odoo.exceptions import UserError
 from odoo.addons.base.models.res_bank import sanitize_account_number
 
-import logging
-_logger = logging.getLogger(__name__)
-
-
-class AccountBankStatementLine(models.Model):
-    _inherit = "account.bank.statement.line"
-
-    # Ensure transactions can be imported only once (if the import format provides unique transaction ids)
-    unique_import_id = fields.Char(string='Import ID', readonly=True, copy=False)
-
-    _sql_constraints = [
-        ('unique_import_id', 'unique (unique_import_id)', 'A bank account transactions can be imported only once !')
-    ]
-
 
 class AccountBankStatementImport(models.TransientModel):
     _name = 'account.bank.statement.import'
