+++ addons/account_bank_statement_import/models/account_bank_statement_line.py
@@ -1,0 +1,12 @@
+from odoo import fields, models
+
+
+class AccountBankStatementLine(models.Model):
+    _inherit = "account.bank.statement.line"
+
+    # Ensure transactions can be imported only once (if the import format provides unique transaction ids)
+    unique_import_id = fields.Char(string='Import ID', readonly=True, copy=False)
+
+    _sql_constraints = [
+        ('unique_import_id', 'unique (unique_import_id)', 'A bank account transactions can be imported only once !')
+    ]
